pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
log_level = "info"
log_cli = true
log_cli_level = "info"
log_format = "%(asctime)s %(name)s %(levelname)s %(message)s"
log_date_format="%Y-%m-%d %H:%M:%S"
filterwarnings = [
    'ignore::DeprecationWarning',
]
testpaths = [
    "tests",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[project]
name = "rl-wizz"
version = "0.1.0"
description = ""
authors = [
    {name = "Javier Smith",email = "javier.smith.dlc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "streamlit (>=1.43.2,<2.0.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "langgraph (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.10,<0.4.0)",
    "streamlit-extras (>=0.6.0,<0.7.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "langchain-pinecone (>=0.2.3,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
